progress, see what breaks
