+++ addon/-private/task-properties.js
@@ -6,11 +6,6 @@ import EnqueueSchedulerPolicy from './external/scheduler/policies/enqueued-polic
 import DropSchedulerPolicy from './external/scheduler/policies/drop-policy';
 import KeepLatestSchedulerPolicy from './external/scheduler/policies/keep-latest-policy';
 import RestartableSchedulerPolicy from './external/scheduler/policies/restartable-policy';
-import {
-  task as taskDecorator,
-  taskGroup as taskGroupDecorator,
-} from './task-decorators';
-import { TaskFactory } from './task-factory';
 
 let taskFactorySymbol = '__ec_task_factory';
 
@@ -182,25 +177,6 @@ export const propertyModifiers = {
   },
 };
 
-function isDecoratorOptions(possibleOptions) {
-  if (!possibleOptions) {
-    return false;
-  }
-  if (typeof possibleOptions === 'function') {
-    return false;
-  }
-
-  if (
-    typeof possibleOptions === 'object' &&
-    'perform' in possibleOptions &&
-    typeof possibleOptions.perform === 'function'
-  ) {
-    return false;
-  }
-
-  return Object.getPrototypeOf(possibleOptions) === Object.prototype;
-}
-
 /**
   A {@link TaskProperty} is the Computed Property-like object returned
   from the {@linkcode task} function. You can call Task Modifier methods
@@ -331,109 +307,5 @@ export function taskComputed(fn) {
     return cp;
   } else {
     return computed(fn);
-  }
-}
-
-/**
- * A Task is a cancelable, restartable, asynchronous operation that
- * is driven by a generator function. Tasks are automatically canceled
- * when the object they live on is destroyed (e.g. a Component
- * is unrendered).
- *
- * To define a task, use the `task(...)` function, and pass in
- * a generator function, which will be invoked when the task
- * is performed. The reason generator functions are used is
- * that they (like the proposed ES7 async-await syntax) can
- * be used to elegantly express asynchronous, cancelable
- * operations.
- *
- * You can also define an
- * <a href="/docs/advanced/encapsulated-task">Encapsulated Task</a>
- * by passing in an object that defined a `perform` generator
- * function property.
- *
- * The following Component defines a task called `myTask` that,
- * when performed, prints a message to the console, sleeps for 1 second,
- * prints a final message to the console, and then completes.
- *
- * ```js
- * import { task, timeout } from 'ember-concurrency';
- * export default Component.extend({
- *   myTask: task(function * () {
- *     console.log("Pausing for a second...");
- *     yield timeout(1000);
- *     console.log("Done!");
- *   })
- * });
- * ```
- *
- * ```hbs
- * <button {{action myTask.perform}}>Perform Task</button>
- * ```
- *
- * By default, tasks have no concurrency constraints
- * (multiple instances of a task can be running at the same time)
- * but much of a power of tasks lies in proper usage of Task Modifiers
- * that you can apply to a task.
- *
- * @param {function} generatorFunction the generator function backing the task.
- * @returns {TaskProperty}
- */
-export function task(taskFnOrProtoOrDecoratorOptions, key, descriptor) {
-  if (
-    isDecoratorOptions(taskFnOrProtoOrDecoratorOptions) ||
-    (key && descriptor)
-  ) {
-    return taskDecorator(...arguments);
-  } else {
-    let tp = taskComputed(function () {
-      tp[taskFactorySymbol].setTaskDefinition(tp.taskFn);
-      return tp[taskFactorySymbol].createTask(this);
-    });
-
-    tp.taskFn = taskFnOrProtoOrDecoratorOptions;
-    tp[taskFactorySymbol] = new TaskFactory();
-
-    Object.setPrototypeOf(tp, TaskProperty.prototype);
-
-    return tp;
-  }
-}
-
-/**
- * "Task Groups" provide a means for applying
- * task modifiers to groups of tasks. Once a {@linkcode Task} is declared
- * as part of a group task, modifiers like `drop` or `restartable`
- * will no longer affect the individual `Task`. Instead those
- * modifiers can be applied to the entire group.
- *
- * ```js
- * import { task, taskGroup } from 'ember-concurrency';
- *
- * export default class MyController extends Controller {
- *   &#64;taskGroup({ drop: true }) chores;
- *
- *   &#64;task({ group: 'chores' }) mowLawn = taskFn;
- *   &#64;task({ group: 'chores' }) doDishes = taskFn;
- *   &#64;task({ group: 'chores' }) changeDiapers = taskFn;
- * }
- * ```
- *
- * @returns {TaskGroup}
- */
-export function taskGroup(possibleDecoratorOptions, key, descriptor) {
-  if (isDecoratorOptions(possibleDecoratorOptions) || (key && descriptor)) {
-    return taskGroupDecorator(...arguments);
-  } else {
-    let tp = taskComputed(function (key) {
-      tp[taskFactorySymbol].setName(key);
-      return tp[taskFactorySymbol].createTaskGroup(this);
-    });
-
-    tp[taskFactorySymbol] = new TaskFactory();
-
-    Object.setPrototypeOf(tp, TaskGroupProperty.prototype);
-
-    return tp;
   }
 }
