+++ addon/index.js
@@ -2,9 +2,11 @@ import { timeout, EmberYieldable as Yieldable } from './-private/utils';
 import {
   TaskProperty,
   TaskGroupProperty,
+} from './-private/task-properties';
+import {
   task,
   taskGroup,
-} from './-private/task-properties';
+} from './-private/task-public-api';
 import { TaskInstance } from './-private/task-instance';
 import {
   all,
